Refactor: Add Store.environment
