+++ core/jvm/src/test/scala/japgolly/clearconfig/JvmTest.scala
@@ -1,6 +1,5 @@
 package japgolly.clearconfig
 
-import cats.syntax.apply._
 import cats.{Eq, Id}
 import japgolly.microlibs.testutil.TestUtil._
 import java.io.File
@@ -11,11 +10,6 @@ import utest._
 object JvmTest extends TestSuite {
 
   override def tests = Tests {
-
-    "environment" - {
-      val s = ConfigSource.environment[Id]
-      println(ConfigDef.get[String]("JAVA.HOME").withReport.run(s).getOrDie()._2.obfuscateKeys(_.value.toUpperCase.contains("PASS")).full)
-    }
 
     "propFileOnClasspath" - {
 
@@ -61,50 +55,6 @@ object JvmTest extends TestSuite {
         assertEq(r, ConfigResult.Success(None))(Eq.fromUniversalEquals, implicitly)
       }
 
-    }
-
-    "inlineProps" - {
-
-      "ok" - {
-        val src1 = ConfigSource.manual[Id]("a")(
-          "x.1" -> "hehe",
-          "INLINE" ->
-            s"""
-               |# hehe
-               | x.2      = 123
-               | x.3    = good stuff # nice
-               |""".stripMargin
-        )
-        val src = ConfigSource.expandInlineProperties(src1, "INLINE")
-
-        val cfgDef = (
-          ConfigDef.need[String]("x.1"),
-            ConfigDef.need[String]("x.2"),
-            ConfigDef.need[String]("x.3")
-          ).tupled.withReport
-
-        val (xs, report) = cfgDef.run(src).getOrDie()
-        assert(xs == (("hehe", "123", "good stuff")))
-        assert(!report.full.contains("INLINE"))
-        report.full
-      }
-
-      "ko" - {
-        val src1 = ConfigSource.manual[Id]("a")(
-          "x.1" -> "hehe",
-          "INLINE" ->
-            s"""
-               |# hehe
-               | x.1      = 123
-               |""".stripMargin
-        )
-        val src = ConfigSource.expandInlineProperties(src1, "INLINE")
-
-        val cfgDef = ConfigDef.need[String]("x.1")
-
-        val result = cfgDef.run(src).toEither
-        assert(result == Left("Error preparing source [SourceName(a)]: The following keys are defined at both the top-level and in INLINE: x.1."))
-      }
     }
 
   }
